project2/requirements.txt
fastapi==0.111.0
langchain==0.2.4
langchain_community==0.2.4
langchain_core==0.2.6
langchain_openai==0.1.8
langgraph==0.0.68
langserve==0.2.2
numpy==1.26.4
openai==1.34.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.1
